feat(auth): refactor attestation and add for auth
