+++ rtc_auth_enclave/Cargo.toml
@@ -19,3 +19,7 @@ sgx_tstd = { git = "https://github.com/apache/teaclave-sgx-sdk.git", features = 
 rtc_types = { path = "../rtc_types", features = ["teaclave_sgx"]}
 
 [dependencies]
+rtc_tenclave = { path = "../rtc_tenclave" }
+
+[patch."https://github.com/apache/teaclave-sgx-sdk.git"]
+sgx_types = { git = "https://github.com/apache/incubator-teaclave-sgx-sdk.git" }
