+++ rtc_auth_enclave/src/lib.rs
@@ -8,12 +8,4 @@ use sgx_types::{sgx_report_t, sgx_status_t, sgx_target_info_t};
 
 use rtc_types::{CreateReportResult, EnclaveHeldData};
 
-/// TODO: Stubbed out, for now.
-#[no_mangle]
-pub extern "C" fn enclave_create_report(
-    _p_qe3_target: *const sgx_target_info_t,
-    _enclave_pubkey: *mut EnclaveHeldData,
-    _p_report: *mut sgx_report_t,
-) -> CreateReportResult {
-    CreateReportResult::Sgx(sgx_status_t::SGX_ERROR_UNEXPECTED)
-}
+use rtc_tenclave::enclave::*;
