+++ rtc_tenclave/Cargo.toml
@@ -11,7 +11,7 @@ doctest = false
 crate-type = ["lib"]
 
 [features]
-default = ["sgx_tstd", "sgx_tse", "rtc_types/teaclave_sgx", "rand", "thiserror"]
+default = ["sgx_tstd", "sgx_tse", "rtc_types/teaclave_sgx", "rand", "thiserror", "sgx_tcrypto"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -21,6 +21,7 @@ sgx_tstd = { git = "https://github.com/apache/teaclave-sgx-sdk.git", features = 
 sgx_tse = { git = "https://github.com/apache/teaclave-sgx-sdk.git" , rev = "v1.1.3", optional = true }
 rand = { git = "https://github.com/mesalock-linux/rand-sgx", tag = "v0.7.3_sgx1.1.3", optional = true }
 thiserror = { git = "https://github.com/mesalock-linux/thiserror-sgx.git", optional = true }
+sgx_tcrypto = { git = "https://github.com/apache/teaclave-sgx-sdk.git", optional = true }
 
 rtc_types = { path = "../rtc_types" }
 sgx_types = { git = "https://github.com/apache/teaclave-sgx-sdk.git", features = ["extra_traits"]}
@@ -36,6 +37,7 @@ cfg-if = "1.0.0"
 [dev-dependencies]
 thiserror_std = { package = "thiserror", version = "1.0.9" }
 rand_std = { package = "rand", version = "0.7.3" }
+sgx_ucrypto = { git = "https://github.com/apache/teaclave-sgx-sdk.git" }
 
 
 [patch."https://github.com/apache/teaclave-sgx-sdk.git"]
